Initial Commit
This provides an incomplete implementation of ARB_fragment_program.
Particularly a large number of the PARAM and TYPE cases are not covered,
and some errors are not flagged that should be.

This further does not include even basic compatibility for
ARB_vertex_program, and a good number of the basic defines are the same
and should be put in a shared file and reused.